修改

+++ Lg.EducationPlatform.DAL/BaseDAL.cs
@@ -46,7 +46,7 @@ namespace Lg.EducationPlatform.DAL
         /// </summary>
         /// <param name="id">主键</param>
         /// <returns>受影响的行数</returns>
-        public int Delete(int id)
+        public int Delete(long id)
         {
             T model = dbContext.Set<T>().Find(id);
             DbEntityEntry<T> entry = dbContext.Entry<T>(model);
@@ -156,7 +156,7 @@ namespace Lg.EducationPlatform.DAL
         /// </summary>
         /// <param name="id">主键</param>
         /// <returns>如果未在上下文或数据源中找到该实体,则返回null</returns>
-        public T GetEntity(int id)
+        public T GetEntity(long id)
         {
             return dbContext.Set<T>().Find(id);
         }
